+++ src/vw/BundleAdjustment/AdjustSparse.h
@@ -123,11 +123,11 @@ namespace ba {
       VW_DEBUG_ASSERT(this->m_control_net->size() == this->m_model.num_points(), LogicErr() << "BundleAdjustment::update() : Number of bundles does not match the number of points in the bundle adjustment model.");
 
       // Reseting the values for U, V, epsilon
-      for ( uint32 j = 0; j < this->m_model.num_cameras(); j++ ) {
+      for ( size_t j = 0; j < this->m_model.num_cameras(); j++ ) {
         U[j] = matrix_camera_camera();
         epsilon_a[j] = vector_camera();
       }
-      for ( uint32 i = 0; i < this->m_model.num_points(); i++ ) {
+      for ( size_t i = 0; i < this->m_model.num_points(); i++ ) {
         V[i] = matrix_point_point();
         epsilon_b[i] = vector_point();
       }
@@ -140,18 +140,19 @@ namespace ba {
       // matrix.
       time.reset(new Timer("Solve for Image Error, Jacobian, U, V, and W:", DebugMessage, "ba"));
       double error_total = 0; // assume this is r^T\Sigma^{-1}r
-      for ( uint32 j = 0; j < m_crn.size(); j++ ) {
+      for ( size_t j = 0; j < m_crn.size(); j++ ) {
         for ( crn_iter fiter = m_crn[j].begin();
               fiter != m_crn[j].end(); fiter++ ) {
-          uint32 i = (**fiter).m_point_id;
+          size_t i = (**fiter).m_point_id;
 
-          matrix_2_camera A = this->m_model.A_jacobian( i, j,
-                                                        this->m_model.A_parameters(j),
-                                                        this->m_model.B_parameters(i) );
-
-          matrix_2_point B = this->m_model.B_jacobian( i, j,
-                                                       this->m_model.A_parameters(j),
-                                                       this->m_model.B_parameters(i) );
+          matrix_2_camera A =
+            this->m_model.A_jacobian( i, j,
+                                      this->m_model.A_parameters(j),
+                                      this->m_model.B_parameters(i) );
+          matrix_2_point B =
+            this->m_model.B_jacobian( i, j,
+                                      this->m_model.A_parameters(j),
+                                      this->m_model.B_parameters(i) );
 
           // Apply robust cost function weighting
           Vector2 error;
@@ -238,7 +239,7 @@ namespace ba {
         matrix_camera_camera u_lambda;
         u_lambda.set_identity();
         u_lambda *= this->m_lambda;
-        for ( uint32 i = 0; i < U.size(); ++i )
+        for ( size_t i = 0; i < U.size(); ++i )
           U[i] += u_lambda;
       }
 
@@ -246,7 +247,7 @@ namespace ba {
         matrix_point_point v_lambda;
         v_lambda.set_identity();
         v_lambda *= this->m_lambda;
-        for ( uint32 i = 0; i < V.size(); ++i )
+        for ( size_t i = 0; i < V.size(); ++i )
           V[i] += v_lambda;
       }
       time.reset();
@@ -262,14 +263,14 @@ namespace ba {
       }
 
       // Compute V inverse
-      for ( uint32 i = 0; i < this->m_model.num_points(); i++ ) {
+      for ( size_t i = 0; i < this->m_model.num_points(); i++ ) {
         Matrix<double> V_temp = V[i];
         chol_inverse( V_temp );
         V_inverse[i] = transpose(V_temp)*V_temp;
       }
 
       // Compute Y and finish constructing e.
-      for ( uint32 j = 0; j < m_crn.size(); j++ ) {
+      for ( size_t j = 0; j < m_crn.size(); j++ ) {
         for ( crn_iter fiter = m_crn[j].begin();
               fiter != m_crn[j].end(); fiter++ ) {
           // Compute the blocks of Y
@@ -292,7 +293,7 @@ namespace ba {
       // below.
       math::MatrixSparseSkyline<double> S(this->m_model.num_cameras()*num_cam_params,
                                           this->m_model.num_cameras()*num_cam_params);
-      for ( uint32 j = 0; j < m_crn.size(); j++ ) {
+      for ( size_t j = 0; j < m_crn.size(); j++ ) {
         { // Filling in diagonal
           matrix_camera_camera S_jj;
 
@@ -306,19 +307,19 @@ namespace ba {
           S_jj += U[j];
 
           // Loading into sparse matrix
-          uint32 offset = j * num_cam_params;
-          for ( uint32 aa = 0; aa < num_cam_params; aa++ ) {
-            for ( uint32 bb = aa; bb < num_cam_params; bb++ ) {
+          size_t offset = j * num_cam_params;
+          for ( size_t aa = 0; aa < num_cam_params; aa++ ) {
+            for ( size_t bb = aa; bb < num_cam_params; bb++ ) {
               S( offset+bb, offset+aa ) = S_jj(aa,bb);  // Transposing
             }
           }
         }
 
         // Filling in off diagonal
-        for ( uint32 k = j+1; k < m_crn.size(); k++ ) {
+        for ( size_t k = j+1; k < m_crn.size(); k++ ) {
           typedef boost::weak_ptr<JFeature> w_ptr;
           typedef boost::shared_ptr<JFeature> f_ptr;
-          typedef std::multimap< uint32, f_ptr >::iterator mm_iterator;
+          typedef std::multimap< size_t, f_ptr >::iterator mm_iterator;
           std::pair< mm_iterator, mm_iterator > feature_range;
           feature_range = m_crn[j].map.equal_range( k );
 
@@ -380,7 +381,7 @@ namespace ba {
 
         // Building right half, sum( WijT * delta_aj )
         std::vector< vector_point > right_delta_b( this->m_model.num_points() );
-        for ( uint32 j = 0; j < m_crn.size(); j++ ) {
+        for ( size_t j = 0; j < m_crn.size(); j++ ) {
           for ( crn_iter fiter = m_crn[j].begin();
                 fiter != m_crn[j].end(); fiter++ ) {
             right_delta_b[ (**fiter).m_point_id ] += transpose( (**fiter).m_w ) *
@@ -389,7 +390,7 @@ namespace ba {
         }
 
         // Solving for delta b
-        for ( uint32 i = 0; i < this->m_model.num_points(); i++ ) {
+        for ( size_t i = 0; i < this->m_model.num_points(); i++ ) {
           Vector<double> delta_temp = epsilon_b[i] - right_delta_b[i];
           Matrix<double> hessian = V[i];
           solve( delta_temp, hessian );
@@ -400,11 +401,11 @@ namespace ba {
 
       //Predicted improvement for Fletcher modification
       double dS = 0;
-      for ( uint32 j = 0; j < this->m_model.num_cameras(); j++ )
+      for ( size_t j = 0; j < this->m_model.num_cameras(); j++ )
         dS += transpose(subvector(delta_a,j*num_cam_params,num_cam_params))
           * ( this->m_lambda * subvector(delta_a,j*num_cam_params,num_cam_params) +
               epsilon_a[j] );
-      for ( uint32 i = 0; i < this->m_model.num_points(); i++ )
+      for ( size_t i = 0; i < this->m_model.num_points(); i++ )
         dS += transpose(subvector(delta_b,i*num_pt_params,num_pt_params))
           * ( this->m_lambda * subvector(delta_b,i*num_pt_params,num_pt_params) +
               epsilon_b[i] );
@@ -415,7 +416,7 @@ namespace ba {
       // -------------------------------
       time.reset(new Timer("Solve for Updated Error", DebugMessage, "ba"));
       double new_error_total = 0;
-      for ( uint32 j = 0; j < m_crn.size(); j++ ) {
+      for ( size_t j = 0; j < m_crn.size(); j++ ) {
         for ( crn_iter fiter = m_crn[j].begin();
               fiter != m_crn[j].end(); fiter++ ) {
           // Compute error vector
